Switched to codeigniter. Patient access and reading now WORKS.
So in knowing that we are behind and need to get to testers and make a
semi-polished product that we can demo to the class, I have made the
executive decision to use a php framework I am very comfortable with to
get the job done and speed up this process. Julian and Fausto, you guys
did a great job with the database and I would hate for that to go
waste. **We can still use basic PHP, if we want**, but this framework
http://www.codeigniter.com speeds up development by auto-escaping
queries, making accessing post data very easy, and essentially taking a
lot of the pain out of basic php.

I got the view patients section up and running, you can drill into a
patient's data, alter it, and it will save. We need to have patient
history save, Fausto, that is something you can do tomorrow. The way
the framework is structured seems a little strange at first, but trust
me, its helpful. Its a model, view, controller framework. For speed, we
will be breaking the rule and using just the controllers for all
database access and logic, and one single view (2 counting login) for
the site (the rest remains the same, with the templates being loaded by
javascript).

One thing, though, the address to anything but the base controller,
which can be found at the root directory, needs to start with
index.php/, for example, when calling the controller dataAccess from
controllers.js, I use index.php/dataAccess. Ask me why this is later.

Take a quick look at the codeigniter web site, and take a look at my
code to see how it works. Fausto and Julian, you guys will be dealing
mostly with Active Record Class (under the Database class section of
the codeingiter user guide, on their web site) and the file
application/controllers/dataAccess.php.  Just look at what I have there
and it should be mildly obvious.

We don't need to worry about encryption, that's built in, or config, I
already set it up. Just make sure that under
application/config/config.php the base url lines up with your
localhost:PORT/directory. Also, in application/config/database.php,
change the database location and credentials to match your computer.

Other than that, we need to make functions to get and set other pages,
so work on that. Next for me is the login login and session storage,
which I am happy to take on myself.

Again, I wouldn't have changed it up if we didn't absolutely need to
speed the hell up. Doing it this way, I can rapidly develop and help
you guys, and we can get done very quickly. Just spend half an hour on
the web site to understand how it works in general, then just copy my
code in dataAccess to write the other getter and setter methods.

I'm commiting these changes. I have a backup in case you guys
absolutely hate this idea, but I know its for the best, and we will get
more done in less time.

**We can still use basic PHP, if we want, in conjunction with
Codeigniter**